+++ packages/core/src/agent/type.ts
@@ -1,4 +1,4 @@
-import type { BaseEvent } from "../callbacks/CallbackManager.js";
+import type { BaseEvent } from "../internal/type.js";
 
 export type AgentStartEvent = BaseEvent<{}>;
 export type AgentEndEvent = BaseEvent<{}>;
