+++ packages/core/src/tools/index.ts
@@ -1,4 +1,3 @@
 export * from "./QueryEngineTool.js";
 export * from "./WikipediaTool.js";
 export * from "./functionTool.js";
-export * from "./types.js";
